changes : # remove clock bean and update tests to use dynamic dates # remove supper class instead use kotlin extension function
