+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/service/ContactService.kt
@@ -6,14 +6,14 @@ import org.slf4j.LoggerFactory
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.mapping.ContactServiceMappers
+import uk.gov.justice.digital.hmpps.hmppscontactsapi.mapping.mapRelationShip
+import uk.gov.justice.digital.hmpps.hmppscontactsapi.mapping.toModel
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.ContactSearchRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.CreateContactRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.response.Contact
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.repository.ContactRepository
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.repository.ContactSearchRepository
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.repository.PrisonerContactRepository
-import java.time.Clock
 import kotlin.jvm.optionals.getOrNull
 
 @Service
@@ -21,9 +21,8 @@ class ContactService(
   private val contactRepository: ContactRepository,
   private val prisonerContactRepository: PrisonerContactRepository,
   private val prisonerService: PrisonerService,
-  clock: Clock,
   val contactSearchRepository: ContactSearchRepository,
-) : ContactServiceMappers(clock) {
+) {
   companion object {
     private val logger = LoggerFactory.getLogger(this::class.java)
   }
@@ -31,10 +30,9 @@ class ContactService(
   @Transactional
   fun createContact(request: CreateContactRequest): Contact {
     validate(request)
-    val newContact = mapContact(request)
-    val createdContact = mapEntityToContact(contactRepository.saveAndFlush(newContact))
-    val newRelationship = request.relationship
-      ?.let { mapRelationShip(createdContact, request.relationship, request) }
+    val newContact = request.toModel()
+    val createdContact = contactRepository.saveAndFlush(newContact).toModel()
+    val newRelationship = request.mapRelationShip(createdContact)
       ?.let { prisonerContactRepository.saveAndFlush(it) }
 
     logger.info("Created new contact {}", createdContact)
@@ -44,14 +42,11 @@ class ContactService(
 
   fun getContact(id: Long): Contact? {
     return contactRepository.findById(id).getOrNull()
-      ?.let { entity -> mapEntityToContact(entity) }
+      ?.toModel()
   }
 
   fun searchContacts(pageable: Pageable, request: ContactSearchRequest): Page<Contact> =
-    contactSearchRepository.searchContacts(request, pageable)
-      .map { entity ->
-        mapEntityToContact(entity)
-      }
+    contactSearchRepository.searchContacts(request, pageable).toModel()
 
   private fun validate(request: CreateContactRequest) {
     if (request.relationship != null) {
